Basic Search Functionality down
Logged in user can now loosely search for individual photos among their
collection. There are still a few kinks to work out of course but they
main functionality is there. A big problem is that once the user hits
search the user is kicked back to the ‘recent’ tab instead of the
‘search’ tab to view their results. I will be working to fix this
annoying problem. But for now I will close this issue. fix #9