User's can now create albums
Right now, the user flow is a bit off, because the user HAS to create
the album first, but  everything in the database should be set up
correctly. I link each photo to the album id they are in. Now I am
going to work on displaying the albums with all their respective photos
inside. Working on that. Fix #13