Rudimentary Sort implemented
Added form on the feed page that changes how you wish to sort your
albums. It’s very basic, I plan to add more to it obviously but since I
have the basic framework of it I am going to keep moving forward and
finishing the other main functionality. that was my fifth issue for
this milestone. fix #10