User can now delete multiple photos
I need to fix the album total  photo count, because it does not
subtract yet after deleting multiple photos, but the user can now
delete multiple pictures from an album at once. Of course this needs to
be ironed out and perfected more but for now it is done. Fix #15